A generalized move capture in lambda with a return object of wrraped Event-Task. usage: https://github.com/usi-verification-and-security/SMTS/blob/dc4c26728acb0645cd5f4cfb02066511c6a78bf2/src/lemma_server/LemmaServer.cpp#L137
implementation of a functor which will capture std::move(p) as it is passed to capture(). The second argument of capture is a lambda which takes the captured variable as an argument. When lambda is used as a function object, then all arguments that are passed to it will be forwarded to the internal lambda as arguments after the captured variable.